+++ hal_psee_plugins/include/devices/gen31/gen31_evk2_tz_device.h
@@ -41,7 +41,7 @@ public:
     virtual I_HW_Identification::SensorInfo get_sensor_info() {
         return {3, 1, "Gen3.1"};
     }
-    long long get_sensor_id();
+    int64_t get_sensor_id();
     virtual int get_temperature();
     virtual int get_illumination();
