+++ hal_psee_plugins/include/devices/imx636/imx636_evk2_esp_tz_device.h
@@ -44,7 +44,7 @@ public:
     virtual I_HW_Identification::SensorInfo get_sensor_info() {
         return {4, 2, "IMX636"};
     }
-    long long get_sensor_id();
+    int64_t get_sensor_id();
     virtual int get_illumination();
     virtual int get_temperature();
